app.js의 주석처리된 라우팅 제거, authRoute리팩토링, 유효성검사부분 미들웨어화
