+++ app.js
@@ -7,8 +7,6 @@ const authRouter = require("./routes/authRoutes");
 const port = process.env.PORT;
 const app = express();
 
-//const authenticateJWT = require('./middlewares/authmiddleware');
-
 app.use(bodyParser.json());
 app.use(express.urlencoded({ extended: true }));
 app.use('/auth', authRouter);
@@ -16,10 +14,6 @@ app.use('/auth', authRouter);
 app.get('/', (req, res) => {
     res.send("hello");
 });
-
-// app.get('/protected', authenticateJWT, (req, res) => {
-//     res.json({ message: "Login Success!" });
-// });
 
 app.listen(port, () => {
     console.log(`Server is listening on ${port}`);
