+++ middlewares/registerMiddleware.js
@@ -1,0 +1,11 @@
+const { validationResult } = require('express-validator');
+
+const validateRegister = (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+    }
+    next();
+};
+
+module.exports = validateRegister;
